trusty: Avoid signed overflow
SMC_FC_FFA_MEM_FRAG_RX returns the 16 bit sender-id in the top 16 bits
of 32 bit register. Cast to uint32_t before shifting it up to avoid
having the uint16_t promoted to a signed int for the shift.

Currently the value we return will never have the top bit of the
sender-id in this context, but this could change in the future.

Bug: 185307117

+++ services/spd/trusty/shared-mem-smcall.c
@@ -317,7 +317,7 @@ static long trusty_ffa_fill_desc(struct trusty_shmem_client_state *client,
 	if (obj->desc_filled != obj->desc_size) {
 		SMC_RET8(smc_handle, SMC_FC_FFA_MEM_FRAG_RX, handle_low,
 			 handle_high, obj->desc_filled,
-			 obj->desc.sender_id << 16, 0, 0, 0);
+			 (uint32_t)obj->desc.sender_id << 16, 0, 0, 0);
 	}
 
 	SMC_RET8(smc_handle, SMC_FC_FFA_SUCCESS, 0, handle_low, handle_high, 0,
